src/backend/requirements.txt
azure-functions
fastapi
openai
